pyproject.toml
[project]
name = "coding-env"
version = "0.1.0"
description = "Coding environment and ast-pilot generator for 0-to-1 development tasks"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "hud-python[agents]>=0.5.35",
    "rubric>=1.3,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff==0.11.2",
    "pre-commit>=3.5.0",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
ast-pilot = "ast_pilot.cli:main"
hud_eval = "cli:main"

[tool.uv]
package = true

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.hud]
image = "01-coding-template:dev"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "C4", "UP"]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/ast_pilot", "."]
